// resources/react/interfaces/ICarRequestFormData.ts
export interface ICarRequestFormData {
    fullName: string;
    phone: string;
    city: string;
    employerType: string;
    yearsOfService: number;
    salary: string;
    obligations: string;
    hasPersonalLoan: boolean;
    hasMortgageLoan: boolean;
    hasSimahDefault: boolean;
    hasTrafficViolations: boolean;
}